Create animation module, replace .map() with regular loop
