+++ lib/anime.umd.js
@@ -836,31 +836,35 @@
     });
   }
 
-  // Animations
-
   function createAnimation(animatable, prop) {
     const animType = getAnimationType(animatable.target, prop.name);
     if (animType) {
       const tweens = normalizeTweens(prop, animatable);
+      const firstTween = tweens[0];
       const lastTween = tweens[tweens.length - 1];
       return {
         type: animType,
         property: prop.name,
         animatable: animatable,
         tweens: tweens,
+        delay: firstTween.delay,
         duration: lastTween.end,
-        delay: tweens[0].delay,
         endDelay: lastTween.endDelay
       }
     }
   }
 
   function getAnimations(animatables, properties) {
-    return filterArray(flattenArray(animatables.map(animatable => {
-      return properties.map(prop => {
-        return createAnimation(animatable, prop);
-      });
-    })), a => !is.und(a));
+    const animations = [];
+    for (let a = 0, aLength = animatables.length; a < aLength; a++) {
+      const animatable = animatables[a];
+      if (animatable) {
+        for (let p = 0, pLength = properties.length; p < pLength; p++) {
+          animations.push(createAnimation(animatable, properties[p]));
+        }
+      }
+    }
+    return animations;
   }
 
   // Create Instance
